+++ FrontEnd/assets/style.css
@@ -243,10 +243,7 @@ li:hover {
 .update-portfolio-hide {
   display: none;
 }
-/* #portfolio h2 {
-	text-align: center;
-	margin-bottom: 1em;
-} */
+
 .gallery {
   width: 100%;
   display: grid;
@@ -259,7 +256,6 @@ li:hover {
   object-fit: cover;
   width: 100%;
   height: 500px;
-  /* height: auto; */
 }
 #contact {
   width: 50%;
@@ -351,6 +347,8 @@ footer nav ul {
   visibility: hidden;
 }
 
+/*Manage the modals content*/
+
 .modal {
   display: none;
   position: fixed;
@@ -369,7 +367,6 @@ footer nav ul {
   padding: 20px;
   border: 1px solid #888;
   width: 30%;
-  /* height: 75%; */
   height: 700px;
   border-radius: 10px;
   display: flex;
@@ -382,7 +379,6 @@ footer nav ul {
 }
 
 .close-modal {
-  /* align-self: flex-end; */
   color: black;
   font-size: 28px;
   font-weight: bold;
@@ -390,7 +386,6 @@ footer nav ul {
 }
 
 .back-modal {
-  /* align-self: flex-end; */
   color: rgba(0, 0, 0, 1);
   font-size: 28px;
   font-weight: bold;
