+++ FrontEnd/index.html
@@ -21,7 +21,9 @@
     <link rel="stylesheet" href="./assets/style.css" />
     <script src=" ./assets/js/works.js" type="module" defer></script>
     <script src=" ./assets/js/admin.js" type="module" defer></script>
-    <script src=" ./assets/js/modal.js" type="module" defer></script>
+    <script src=" ./assets/js/modalMain.js" type="module" defer></script>
+    <script src=" ./assets/js/modalRemovePhoto.js" type="module" defer></script>
+    <script src=" ./assets/js/modalAddPhoto.js" type="module" defer></script>
   </head>
   <body>
     <header>
